requirements-test.txt
model_mommy
pytest
pytest-django
